Added test suite to avoid creation of orchestrator in each test class. (#786)
* Added test suite to avoid creation of orchestrator in each test class.
